refactor: build default extensions in another rollup config
