+++ tsconfig.rollup-config-default-extensions.json
@@ -1,0 +1,6 @@
+{
+	"extends": "./tsconfig.json",
+	"include": [
+		"./rollup/rollup.default-extensions.ts"
+	]
+}
